Move ignoreIfPrimaryLeaseLoss to local_store.ts
