+++ packages/firestore/src/local/lru_garbage_collector.ts
@@ -24,8 +24,7 @@ import { primitiveComparator } from '../util/misc';
 import { CancelablePromise } from '../util/promise';
 import { SortedMap } from '../util/sorted_map';
 import { SortedSet } from '../util/sorted_set';
-import { ignoreIfPrimaryLeaseLoss } from './indexeddb_persistence';
-import { LocalStore } from './local_store';
+import { ignoreIfPrimaryLeaseLoss, LocalStore } from './local_store';
 import { PersistenceTransaction } from './persistence';
 import { PersistencePromise } from './persistence_promise';
 import { TargetData } from './target_data';
